Hardhat Fixes/Changes (#2704)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request
This PR fixes hardhat headlamps staying on when a person dies, and gives
xenomorphs the ability to slash them and turn them off when they're on
the ground like with flashlights. I also re-did the hardhat code that
was in place and replaced it with code that's in line with the
flashlight.

It's possible I messed something up, but I tested everything locally and
it seems to work fine. Don't tear me apart too hard if something's out
of line or not up to standard.

# Explain why it's good for the game
standardizing the behavior of a light source is awesome........

# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:IowaPotatoFarmer
add: Added the ability for xenomorphs to turn off hardhats on the
ground.
fix: Hardhat headlamps no longer stay on when their wearer dies.
code: Changed hardhat code to be in line with flashlight code.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: harryob <me@harryob.live>