[WIP] Round Results (Internal use stuff) (#2203)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

Round results internal channel is usable again. Plan is to have two
channels: one for in depth data that can be messed with for data and
another for basic information that can be easily readable. Only the in
depth data is currently ready for this just as we test and make sure
everything is running nicely.

# Explain why it's good for the game

To help with balancing we require easily accessible data.

# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

![image](https://user-images.githubusercontent.com/8919187/210452397-86f5e7e9-0fdb-4e25-84cb-23f72a17e7f2.png)

![image](https://user-images.githubusercontent.com/8919187/210452403-8cfd9c39-5faa-4677-9cdd-8add7d5b9852.png)

</details>


# Changelog

:cl: Morrow
server: Introduces round results webhook.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: harryob <me@harryob.live>
Co-authored-by: forest2001 <41653574+realforest2001@users.noreply.github.com>