Plasma pistols and Spike launchers can be wielded, Removed Wrist Scimitar power reduction. (#2750)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

Plasma pistols and Spike launchers can be wielded.

Removed Wrist Scimitar power reduction.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

> Plasma pistols and Spike launchers can be wielded.

These used to be able to wield, but at some point this feature was
removed, possibly accidentally when someone added 'ITEM_PREDATOR' to
flags_item, which overrode and removed the inherent 'TWOHANDED' on all
guns.

Wielding these weapons now increases accuracy and reduces scatter, in
case you need to be extra precise. It was a bit odd that you couldn't
wield one of these before, though it continues to not be necessary as
one-handed use is strong enough.

> Removed Wrist Scimitar power reduction.

This was added pretty much in fear of the Wrist Scim reintroduction
being too strong, and with it being seen now as the weakest primary a
Yautja has access to, it's pretty much entirely unnecessary and makes
Gear Recovery hell due to the high energy consumption of cloaks and the
plasma caster. Giving up your ability to have a far stronger weapon,
like the sword or glaive, is already enough of a drawback to let
Predators reconsider their choices.

I already have a PR on hard dels open, when that one is closed I will
continue with another backend PR.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:
balance: Plasma pistols and Spike launchers can be wielded.
balance: Removed Wrist Scimitar power reduction.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->