Nerfs to the Scout Custom M4RA. (#2783)
# About the pull request

This PR reverts undocumented changes to the M4RA custom scout rifle,
which was recently merged. Namely, the instant wield, the fire rate
delay decrease by 2, a damage multiplier of 0.30, and a accuracy
multiplier as well.

# Explain why it's good for the game

Undocumented, balance changes in a PR that says "with less balance
changes", which severely buffed the scout rifle to be an extremely quick
and safe to say OP weapon. Not only does it have FULL AP (meaning no
castes aside from max rage berserker can have a semblance of reduced
damage), 66 damage on just the base ammo (which is given in abundance),
it also have absolutely 0 fall off (which i did not change, as I believe
scouts shouldn't have falloff but in the case of current scout rifle it
is far too powerful with the other buffed stats). It absolutely shreds
almost every caste, and is bad balancing. This PR reverts those changes.


![image](https://user-images.githubusercontent.com/125149403/223005863-5c233a7e-5a17-4220-a75c-4a74dd16b1a9.png)

This isn't documenting the insane buffs to the scout rifle, and I'd go
so far as to say the author of the PR understood these changes would
make the rifle extremely powerful. Also, the issue with "omega slime"
with ammo switching on scout could be solved without the need for a full
stat increase on damage, accuracy, fire rate.

# Changelog
:cl:
balance: Nerfs to the Scout Custom M4RA.
/:cl:

---------

Co-authored-by: Morrow <darthbane97@gmail.com>