Init Runtime Errors Catcher V2 (#2832)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

There's always been a few issues with logging of early init errors:
* Errors during early /world/New are logged to STUI, but not to runtime
error log file
* Errors during static initializers are not logged anywhere, and even
dodge auxtools debugger since it's attached in /world/New

Addition of SSearlyruntimes was supposed to shed light on these, but it
was too limited and they'd usually get noticed after-the-fact

This does the same thing but baked in /world/New, and ensure errors are
written to both STUI and log files.

The errors in log file are strongly emphased in a block because they
might not get picked up by debugguer. SSearlyruntimes is made into a
simple error message on init.

# Changelog

:cl:
fix: Changed method of logging early game errors to be more convenient.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->