The Shakeship (#2864)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

Admin tool to shake the ship, and a proc that shakes it too when called.

# Explain why it's good for the game

Good tool for events, and shortens some coding, along with future
systems that would cause this shaking.


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:
admin: panel that allows admins to trash the ship about.
code: adds a shakeship proc that can be called to rock the shipmap.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: harryob <me@harryob.live>