Lessens the frequency of Almayer ambient sounds (#2911)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

What it says on the tin: lowers the frequency of ambient sounds and
noises on the Almayer. Should help alleviate otherwise-atmospheric
noises from becoming unbearably annoying.

Also removed a CIC ambient sound that was long and didn't fit well with
the rest of the ambience.

# Explain why it's good for the game

Less earworms! Also should make whenever the ambience _does_ come on a
welcome sight rather than an annoying one.

</details>


# Changelog

:cl:
code: Reduced the frequency of several ambient sounds on the Almayer to
make them less annoying.
sounddel: Removed a CIC ambient sound that didn't fit with the rest.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->